code/requirements.txt
Authlib==0.15.2
Flask==1.1.2
marshmallow==3.9.1
requests==2.25.0
zappa==0.52.0
cryptography==3.3.1
flake8==3.8.4
coverage==5.3
pytest==6.1.2
pyjwt[crypto]==2.0.0